melhorias em carros e marcas

+++ CooperSystem.API/Controllers/MarcasController.cs
@@ -1,16 +1,12 @@
 ﻿using CooperSystem.API.Interfaces.Services;
 using CooperSystem.API.Models;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace CooperSystem.API.Controllers
-{
-    [Route("api/[controller]")]
+{   
     [ApiController]
+    [CultureRoute("api/[controller]")]
     public class MarcasController : ControllerBase
     {
         private readonly IMarcaService _marcaService;
@@ -19,64 +15,76 @@ namespace CooperSystem.API.Controllers
         {
             _marcaService = marcaService;
         }
-
-        [Route("All")]
-        [HttpGet]
-        public IActionResult GetAll()
-        {
-            try
-            {
-                var result = _marcaService.GetAll();
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
-
-        [Route("Filter/{nome}/{origem}")]
-        [HttpGet]
-        public IActionResult GetByFilter(string nome, string origem)
-        {
-            try
-            {
-                var result = _marcaService.GetByFilter(nome, origem);
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
-
-        [Route("Insert")]
+        
+        [Route("Inserir")]
         [HttpPost]
-        public IActionResult Post(Marca marca)
+        public IActionResult Inserir(Marca marca)
         {
             try
             {
-                _marcaService.Insert(marca);
-                return Ok();
+                _marcaService.Inserir(marca);
+                return Ok(new { Message = "Marca adicionada com sucesso!" });
             }
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                return BadRequest(new { Error = "Erro ao adicionar a marca.", Details = ex.Message });
             }
         }
 
-        [Route("Edit")]
+        [Route("Atualizar")]
         [HttpPut]
-        public IActionResult Edit(Marca marca)
+        public IActionResult Atualizar(Marca marca)
         {
             try
             {
-                _marcaService.Edit(marca);
-                return Ok();
+                if (marca.Id == 0)
+                    return BadRequest(new { Error = "id da marca obrigatório" });
+
+                if (marca.Origem == null)
+                    return BadRequest(new { Error = "Origem da marca obrigatório" });
+
+                _marcaService.Atualizar(marca);
+                return Ok(new { Message = "Marca editada com sucesso!" });
             }
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                return BadRequest(new { Error = "Erro ao editar a marca.", Details = ex.Message });
+            }
+        }
+
+        [Route("listarTodasAsMarcas")]
+        [HttpGet]
+        public IActionResult listarTodasAsMarcas()
+        {
+            try
+            {
+                var result = _marcaService.listarTodasAsMarcas();
+                return Ok(new { Message = "Lista de marcas obtida com sucesso!", Data = result });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { Error = "Erro ao obter a lista de marcas.", Details = ex.Message });
+            }
+        }
+
+        [Route("ObterPorNomeEOrigem")]
+        [HttpGet]
+        public IActionResult ObterPorNomeEOrigem(string nome, string origem)
+        {
+            try
+            {
+                if (nome == null)
+                    return BadRequest(new { Error = "Nome da marca obrigatório" });
+
+                if (origem == null)
+                    return BadRequest(new { Error = "Origem da marca obrigatório" });
+
+                var result = _marcaService.ObterPorNomeEOrigem(nome, origem);
+                return Ok(new { Message = "Marcas filtradas com sucesso!", Data = result });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { Error = "Erro ao filtrar as marcas.", Details = ex.Message });
             }
         }
 
@@ -86,13 +94,17 @@ namespace CooperSystem.API.Controllers
         {
             try
             {
+                if (id == 0)
+                    return BadRequest(new { Error = "id da marca obrigatório" });
+
                 _marcaService.Delete(id);
-                return Ok();
+                return Ok(new { Message = "Marca deletada com sucesso!" });
             }
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                return BadRequest(new { Error = "Erro ao deletar a marca.", Details = ex.Message });
             }
         }
+
     }
 }
